adjust file

+++ SQLQuery1.sql
@@ -2,5 +2,5 @@ Create Database rest;
 create table Menu(
 Id int,
 name varchar(50),
-price float
+price decimal
 );
